+++ CHANGES.md
@@ -56,6 +56,7 @@ Changes in 10.0
 - Build command for executing mysqldump in DatabaseBackup.php
 - SQL to extract Unix timestamp or epoch from date in Eligibility/Student.php, StudentList.php & TeacherCompletion.php
 - Install module/plugin: execute the install_mysql.sql script for MySQL in Modules.inc.php, Plugins.inc.php & modules/README.md & plugins/README.md
+- Fix typo "inexistant" to "nonexistent" & update translations in Modules.inc.php, Plugins.inc.php & rosariosis.po
 
 Changes in 9.3
 --------------